+++ src/lib/plugins/links.ts
@@ -32,6 +32,10 @@ export function handleLinks(
     newUrl = handleTSLinks(newUrl);
   }
 
+  const base = dirname.split('/').splice(0, 2).join('/');
+
+  newUrl = replaceInternalLinks(newUrl ?? node.url, base);
+
   return newUrl;
 }
 
@@ -106,12 +110,14 @@ function getNewUrl(node: any, dirname: string) {
     newUrl = newUrl.replace('fuels-ts/guide/', 'fuels-ts/');
   }
 
-  newUrl = newUrl.replace(
-    '/docs/dev/getting-started',
-    isNightly
-      ? '/docs/nightly/wallet/dev/getting-started'
-      : '/docs/wallet/dev/getting-started'
-  );
+  newUrl = newUrl
+    .replace(
+      '/docs/dev/getting-started',
+      isNightly
+        ? '/docs/nightly/wallet/dev/getting-started'
+        : '/docs/wallet/dev/getting-started'
+    )
+    .replace('/api/interfaces/index', '/api/interfaces/');
 
   if (newUrl.includes('/docs/forc/../')) {
     newUrl = newUrl.replace('/docs/forc/../', '/docs/sway/');
@@ -123,6 +129,8 @@ function getNewUrl(node: any, dirname: string) {
       'https://github.com/FuelLabs/fuels-wallet/blob/master'
     );
   }
+
+  console.log(newUrl);
 
   return newUrl;
 }
@@ -217,4 +225,63 @@ function handleNewURLs(
       }
     });
   }
+}
+
+function replaceInternalLinks(href: string, base: string) {
+  if (
+    href.startsWith('https://fuellabs.github.io') &&
+    !href.includes('fuellabs.github.io/block-explorer-v2') &&
+    !href.startsWith('https://fuellabs.github.io/sway/master/std/') &&
+    !href.includes('LICENSE')
+  ) {
+    href = href
+      .replace('https://fuellabs.github.io', '')
+      .replace('/master/', '/')
+      .replace('.html', '')
+      .replace('/nightly', '')
+      .replace(/\/index$/, '/')
+      .replace('sway/book/', 'sway/')
+      .replace('sway/forc/', 'forc/')
+      .replace('/fuel-specs/', '/specs/')
+      .replace(/\/v\d+\.\d+\.\d+\//, '/')
+      .replace('/specs/vm', '/specs/fuel-vm');
+    href = `/docs${href}`;
+
+    const isSwayVersion = href.match(/sway\/(v.+)\/forc/);
+    if (isSwayVersion) {
+      const version = isSwayVersion[1];
+      href = href.replace(`sway/${version}/forc`, 'forc');
+    }
+  }
+
+  if (href.startsWith('../')) {
+    href = href.replace('../', `/${base}/`);
+  }
+  if (href.startsWith('./../')) {
+    href = href.replace('./../', `/${base}/`);
+  }
+
+  if (!href.endsWith('/forc/plugins/forc_client/')) {
+    href = href.replace('/forc/plugins/forc_client/', '/forc/plugins/');
+  }
+
+  // TODO: fix this at source
+  href = href
+    .replace(
+      'docs/fuel-docs/quickstart/developer-quickstart',
+      '/guides/quickstart/'
+    )
+    .replace(
+      'https://fuelbook.fuel.network/master/quickstart/developer-quickstart.html',
+      '/guides/quickstart/'
+    )
+    .replace('specs/fuel-vm/instruction_set', 'specs/fuel-vm/instruction-set')
+    .replace('specs/protocol/tx_format', 'specs/tx-format/')
+    .replace('docs/fuelup/latest', 'docs/fuelup')
+    .replace('specs/protocol/id/contract', 'specs/identifiers/contract-id')
+    .replace('/packag/', '/package/')
+    .replace('/index#', '#');
+
+  console.log('href', href);
+  return href;
 }
